pyproject.toml
[tool.poetry]
name = "cosmic-python-rag-rag"
version = "0.1.0"
description = ""
authors = ["Biszkoptu <filip.jatelnicki@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
scikit-learn = "^1.5.2"
openai = "^1.47.1"
numpy = "1.21.0"
setuptools = "^75.1.0"
pandas = "<2.2.3"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"
nltk = "^3.9.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
